requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
numpy==1.24.3
faiss-cpu==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
backoff==2.2.1
psutil==5.9.6